Words themselves are not anagrams
The word itself is not an anagram.  For instance the word 'AC' has no anagrams but this function would return that it is.  However the word 'AntHill' is a word and the function will return it twice.  There are other areas in this toolset that behave erratically because of this fact and will be corrected by this and others that will not.  The right thing to do is just return none as in the case of 'AC' or one as in the case of 'Anthill'.